Split up the master class into smaller ones
